최종본!

+++ src/main/java/com/cc/model/dto/UserPlayDto.java
@@ -1,13 +1,11 @@
 package com.cc.model.dto;
 
-import javax.persistence.Column;
-
 import lombok.Data;
 
 @Data
 public class UserPlayDto {
 	private String merchantUid;
-	private long userId;
+	private int userId;
 	private String playId;
 	private String playTitle;
 	private String price;
